requirements.txt
flatbuffers==1.11
xlrd==1.2.0
pyinstaller==3.5
pyyaml==5.1.2